+++ log/keyvalue_test.go
@@ -361,6 +361,19 @@ func TestValueFromAttribute(t *testing.T) {
 			v:    attribute.StringSliceValue([]string{"foo", "bar"}),
 			want: log.SliceValue(log.StringValue("foo"), log.StringValue("bar")),
 		},
+		{
+			desc: "Slice",
+			v: attribute.SliceValue([]attribute.Value{
+				attribute.BoolValue(true),
+				attribute.StringValue("foo"),
+				attribute.SliceValue([]attribute.Value{attribute.IntValue(7)}),
+			}),
+			want: log.SliceValue(
+				log.BoolValue(true),
+				log.StringValue("foo"),
+				log.SliceValue(log.Int64Value(7)),
+			),
+		},
 	}
 	for _, tc := range testCases {
 		t.Run(tc.desc, func(t *testing.T) {
@@ -422,6 +435,20 @@ func TestKeyValueFromAttribute(t *testing.T) {
 			desc: "StringSlice",
 			kv:   attribute.StringSlice("k", []string{"foo", "bar"}),
 			want: log.Slice("k", log.StringValue("foo"), log.StringValue("bar")),
+		},
+		{
+			desc: "Slice",
+			kv: attribute.Slice("k", []attribute.Value{
+				attribute.BoolValue(true),
+				attribute.StringValue("foo"),
+				attribute.SliceValue([]attribute.Value{attribute.IntValue(7)}),
+			}),
+			want: log.Slice(
+				"k",
+				log.BoolValue(true),
+				log.StringValue("foo"),
+				log.SliceValue(log.Int64Value(7)),
+			),
 		},
 	}
 	for _, tc := range testCases {